requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
ecdsa==0.15
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
okta-jwt==1.3.5
pyasn1==0.4.8
PyJWT==1.7.1
pymongo==3.10.1
python-dotenv==0.10.5
python-jose==3.1.0
requests==2.22.0
rsa==4.0
six==1.14.0
urllib3==1.25.8
Werkzeug==0.16.0
